#64 The wole pack now attacks the player if one of them is aggroed
Added packID to enemy, when a pack is generated all enemies in that pack have the same ID.

checkAggro method has been rewritten to make the whole pack attack if one of them is attacked.

instead of aggroing when the player when they have lost hp they now aggro the player if they have been hit.